made a ToPlay class to run from to hopefully fix the Player issue
